+++ src/USER-EFF/pair_eff_cut.cpp
@@ -1079,7 +1079,7 @@ void PairEffCut::min_x_set(int /*ignore*/)
 double PairEffCut::memory_usage()
 {
   double bytes = maxeatom * sizeof(double);
-  bytes += maxvatom*6 * sizeof(double);
-  bytes += 2 * nmax * sizeof(double);
+  bytes += (double)maxvatom*6 * sizeof(double);
+  bytes += (double)2 * nmax * sizeof(double);
   return bytes;
 }
